src/main/java/com/springframework/controllers/HomeController.java
package com.springframework.controllers;

/*
 * Created by Rabindra on 8/17/18
 */

import com.springframework.services.RecipeService;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;

import java.util.Optional;

@Slf4j
@Controller
public class HomeController {

    private final RecipeService recipeService;

    public HomeController(RecipeService recipeService) {
        this.recipeService = recipeService;
    }

    @RequestMapping({"","/","/index"})
    public String getIndexPage(Model model){
        log.debug("At Home controller");
        model.addAttribute("recipes",recipeService.getRecipes());
        return "index";
    }
}
